refactored data generation scripts
[skip ci]

Jira ticket: CAMS-421
